Remove history-vim entirely
This commit removes the history-vim and scrollback patches, which makes
the code base a lot smaller and simpler. This patches aren't necessary
anymore because I plan to use tmux in the future, which implements these
features a lot better and also adds other nice features. This way my
terminal does only one thing and does it well in the true unix fashion.